Non-functional attempt to switch to API 3.0
I tried to replicate the old API behavior to create a new campaign and
set its content, but this is no longer supported: I get a 405 upon
POSTing on /campaigns/.../content. It seems I need to create a new template
first, but that one requires a standard plan subscription at minimum.