Newsletter: First create template, then campaign
I think I figured out how to generate custom HTML content: With the new
API, the content first needs to be stored as a template, and then the
campaign can use it. This is implemented here, and I can create a test
campaign against a test account of mine. However, this requires a
"Standard" paid subscription plan at the least. I suspect that's already
what we use for the actual account, because I don't see how that would
work even with API 2.0 otherwise, but I'll confirm with the AOZ team
before pushing this out.